# goblas: optimize Daxpy using SSE2 on amd64