--- a/goblas/daxpy_amd64.go
+++ b/goblas/daxpy_amd64.go
@@ -1,0 +1,8 @@
+// Copyright ©2015 The gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package goblas
+
+func daxpyUnitary(alpha float64, x, y []float64)
+func daxpyInc(alpha float64, x, y []float64, n, incX, incY, ix, iy uintptr)